refactor(serialize): make serialization modes standalone translation units
Split the html5lib #document/HTML serializer, minify, markdown, text and
readability out of tree.c (where they were #included fragment headers) into
their own .c translation units compiled directly into the extension.

Inlining on the hot serialize path is preserved by hoisting the shared
primitives into two static-inline headers both tree.c and the serialize
units include:
  - dom/tree_internal.h: struct th_tree, the arena, copy_input_span,
    need_text and is_void_atom, so no per-text-node cross-TU call appears.
  - serialize/internal.h: the sbuf buffer and its ops, the escape
    formatters, the start/close-tag writers, attribute ordering, the
    charset-meta helpers, foreign_attr_namespaced and the shared
    markdown/text block and whitespace predicates.

serialize_compact and collect_text stay file-local in document.c (their
only caller) to keep their codegen identical to before. No behavior change;
the full suite stays green.
