+++ meson.build
@@ -32,6 +32,11 @@ py.extension_module(
         'src/turbohtml/_c/serialize/escape.c',
         'src/turbohtml/_c/serialize/unescape.c',
         'src/turbohtml/_c/serialize/markup.c',
+        'src/turbohtml/_c/serialize/document.c',
+        'src/turbohtml/_c/serialize/minify.c',
+        'src/turbohtml/_c/serialize/markdown.c',
+        'src/turbohtml/_c/serialize/text.c',
+        'src/turbohtml/_c/serialize/readability.c',
         'src/turbohtml/_c/query/xpath/ast.c',
         'src/turbohtml/_c/query/xpath/lexer.c',
         'src/turbohtml/_c/query/xpath/parser.c',
